Model apply as route
