+++ controllers/apply/index.js
@@ -1,7 +1,32 @@
 'use strict';
 const express = require('express');
-const router = express.Router();
+const createFormRouter = require('./forms/create-form-router');
+const formModel = require('./reaching-communities-form');
 
-router.use('/your-idea', require('./reaching-communities'));
+function initYourIdea() {
+    const router = express.Router();
 
-module.exports = router;
+    router.use((req, res, next) => {
+        res.locals.isBilingual = false;
+        next();
+    });
+
+    router.get('/', (req, res) => {
+        res.render('pages/apply/reaching-communities-startpage', {
+            startUrl: `${req.baseUrl}/1`,
+            form: formModel
+        });
+    });
+
+    const routerWithForm = createFormRouter(router, formModel);
+
+    return routerWithForm;
+}
+
+module.exports = () => {
+    const router = express.Router();
+
+    router.use('/your-idea', initYourIdea());
+
+    return router;
+};
